Move methods from `VectorRegisters` to `VectorRegistersExt` with default implementations and change method order
